requirements.txt
# Core
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0

# Database
supabase==1.0.3
sqlmodel==0.0.11
psycopg2-binary==2.9.9
alembic==1.10.4

# HTTP
httpx==0.23.3
python-multipart==0.0.6

# Development
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-cov==4.0.0

# Code Quality
black==23.3.0
isort==5.12.0
mypy==1.2.0

# Type Hints
typing-extensions==4.5.0
pydantic==1.10.7
